requirements.txt
Django==5.2.10
djangorestframework
psycopg2-binary
celery
redis
pandas
openpyxl
